queues/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1733332318622/work
arxiv==2.4.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1763409923949/work
async-timeout==4.0.3
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
bs4==0.0.2
build==1.4.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
chromadb==1.4.1
click==8.3.1
coloredlogs==15.0.1
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
croniter==6.0.0
dataclasses-json==0.6.7
debugpy @ file:///opt/miniconda3/conda-bld/debugpy_1762421637624/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
distro==1.9.0
docker==7.1.0
durationpy==0.10
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1763918002538/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
fastapi==0.128.0
feedparser==6.0.12
filelock==3.20.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.72.0
groq==0.37.1
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
ipykernel @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_ipykernel_1761567969/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1748711175/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_client_1767954655/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_core_1760643864/work
kubernetes==35.0.0
langchain==1.2.7
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-groq==1.1.1
langchain-ollama==1.0.1
langchain-openai==1.1.7
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langserve==0.3.3
langsmith==0.6.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.15.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
node==2.0.0
numpy==2.2.6
oauthlib==3.3.1
odict==2.0.0
ollama==0.6.1
onnxruntime==1.23.2
openai==2.15.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.2
overrides==7.7.0
packaging==25.0
pandas==2.3.3
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==12.1.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1764950726/work
plumber==2.0.0
portalocker==3.2.0
posthog==5.4.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
propcache==0.4.1
protobuf==6.33.4
psutil @ file:///opt/miniconda3/conda-bld/psutil_1761896322276/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==23.0.0
pybase64==1.4.3
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
PyMuPDF==1.26.7
pypdf==6.6.0
PyPika==0.50.0
pyproject_hooks==1.2.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq @ file:///opt/miniconda3/conda-bld/pyzmq_1762375617494/work
qdrant-client==1.16.2
redis==7.1.1
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.1
rpds-py==0.30.0
rq==2.6.1
sgmllib3k==1.0.0
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
sse-starlette==3.2.0
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.50.0
streamlit==1.53.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.0
tornado @ file:///Users/ec2-user/croot/tornado_1765992646290/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typer==0.21.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1758622279606/work
websocket-client==1.9.0
websockets==16.0
wikipedia==1.4.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zope.component==7.1
zope.deferredimport==6.1
zope.deprecation==6.0
zope.event==6.1
zope.hookable==8.2
zope.interface==8.2
zope.lifecycleevent==6.0
zope.proxy==7.1
zstandard==0.25.0
